packages/deepfreeze-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deepfreeze-server"
version = "2.0.0"
description = "Deepfreeze persistent daemon — REST API, job management, scheduling, and SSE events"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Bret Wortman"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "deepfreeze-core>=2.0.0",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "pydantic>=2.0",
    "elasticsearch8>=8.0.0",
    "sse-starlette>=1.6.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.24.0",
]

[project.scripts]
deepfreeze-server = "deepfreeze_server.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["deepfreeze_server"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
